pyproject.toml
[tool.poetry]
name = "ascii-generator"
version = "0.1.0"
description = ""
authors = ["WoutVossen <vossen.w@hotmail.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
pyarrow = "^19.0.0"
datasets = "^3.2.0"
python-dotenv = "^1.0.1"
modal = "^0.73.45"
llama-cpp-python = "^0.3.7"
transformers = "^4.49.0"
torch = "^2.6.0"
mistralrs = "^0.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
